Deprecate @forrestjs/hooks
