+++ packages/core/.gitignore
@@ -1,0 +1,5 @@
+/node_modules
+package-lock.json
+yarn.lock
+
+
