+++ packages/core/.npmignore
@@ -1,0 +1,5 @@
+/docs
+/tests
+/_src
+/_test
+/CHANGELOG.md
