+++ packages/core/docs/create-hook.md
