+++ packages/core/docs/register-action.md
