+++ packages/core/docs/trace-hook.md
