+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@forrestjs/hooks",
-  "description": "DEPRECATED: see @forrestjs/core",
+  "name": "@forrestjs/core",
+  "description": "ForrestJS is the Javascript plugin syste.",
   "version": "4.2.0-alpha.1",
   "main": "src/index.js",
   "author": "Marco Pegoraro <marco.pegoraro@gmail.com",
@@ -8,7 +8,7 @@
     "url": "https://github.com/forrestjs/forrestjs/issues"
   },
   "dependencies": {
-    "@forrestjs/core": "4.2.0-alpha.1"
+    "@marcopeg/dotted": "^1.12.0"
   },
   "homepage": "https://github.com/forrestjs/forrestjs/tree/master/packages/core#readme",
   "keywords": [
