+++ packages/core/src/constants.js
