+++ packages/core/src/errors.js
