+++ packages/core/src/state.js
