+++ packages/core/test/state.test.js
