+++ packages/service-apollo/package.json
@@ -8,7 +8,7 @@
     "url": "https://github.com/forrestjs/forrestjs/issues"
   },
   "dependencies": {
-    "@forrestjs/hooks": "^4.2.0-alpha.1",
+    "@forrestjs/core": "^4.2.0-alpha.1",
     "apollo-boost": "^0.4.9",
     "cross-fetch": "^3.1.5"
   },
