+++ packages/service-jwt/package.json
@@ -8,7 +8,7 @@
     "url": "https://github.com/forrestjs/forrestjs/issues"
   },
   "dependencies": {
-    "@forrestjs/hooks": "^4.2.0-alpha.1",
+    "@forrestjs/core": "^4.2.0-alpha.1",
     "jsonwebtoken": "^8.5.1"
   },
   "homepage": "https://github.com/forrestjs/forrestjs/tree/master/packages/service-jwt#readme",
