+++ packages/service-logger/package.json
@@ -8,7 +8,7 @@
     "url": "https://github.com/forrestjs/forrestjs/issues"
   },
   "dependencies": {
-    "@forrestjs/hooks": "^4.2.0-alpha.1",
+    "@forrestjs/core": "^4.2.0-alpha.1",
     "winston": "^3.4.0"
   },
   "homepage": "https://github.com/forrestjs/forrestjs/tree/master/packages/service-logger#readme",
